requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
crispy-bootstrap4==2022.1
Django==4.2
django-crispy-forms==2.0
django-filter==23.2
djangorestframework==3.14.0
et-xmlfile==1.1.0
importlib-metadata==6.6.0
Markdown==3.4.3
numpy==1.24.3
openpyxl==3.1.2
pandas==2.0.2
Pillow==9.5.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
zipp==3.15.0
